refactor(ui): reorganise AvailableStorageTable files
