+++ ui/src/app/machines/views/MachineDetails/MachineStorage/AvailableStorageTable/AddPartition/AddPartitionFields/AddPartitionFields.tsx
@@ -1,8 +1,8 @@
 import { Col, Input, Row, Select } from "@canonical/react-components";
 import { useFormikContext } from "formik";
 
+import FilesystemFields from "../../FilesystemFields";
 import type { AddPartitionValues } from "../AddPartition";
-import FilesystemFields from "../FilesystemFields";
 
 import FormikField from "app/base/components/FormikField";
 import type { Machine } from "app/store/machine/types";
